refactor: flatten pages dir
