+++ src/pages/AnnualReport.js
@@ -1,8 +1,8 @@
 import React, { useState } from 'react'
-import { ReportData } from '../../components/annual-report/report-data/ReportData.js'
-import { ReportParameterSelector } from '../../components/report-parameter-selector/index.js'
+import { ReportData } from '../components/annual-report/report-data/ReportData.js'
+import { ReportParameterSelector } from '../components/report-parameter-selector/index.js'
 
-const AnnualReport = () => {
+export const AnnualReport = () => {
     const [reportParameters, setReportParameters] = useState({})
 
     return (
@@ -14,5 +14,3 @@ const AnnualReport = () => {
         </>
     )
 }
-
-export default AnnualReport
