+++ src/pages/index.js
@@ -1,0 +1,2 @@
+export { AnnualReport } from './AnnualReport.js'
+export { Configurations } from './Configurations.js'
